test: service 통합테스트
