+++ src/main/java/com/hhplus/ecommerce/infrastructure/BasketRepository.java
@@ -9,9 +9,9 @@ import java.util.Optional;
 
 public interface BasketRepository extends JpaRepository<Basket, Long> {
 
-    Optional<Basket> findByBasketId(long basketId);
+    Optional<Basket> findByBasketId(Long basketId);
     void deleteByBasketId(Long basketId);
     Basket save(Basket basket);
-    @Query(value = "select b from Basket b where b.basketUser = :userId")
-    Basket getUserBasket(@Param("userId") long userId);
+    @Query(value = "select b from Basket b where b.basketUser.userId = :userId")
+    Basket getUserBasket(@Param("userId") Long userId);
 }
