+++ src/test/java/com/hhplus/ecommerce/business/BasketDetailServiceTest.java
@@ -1,98 +1,99 @@
 package com.hhplus.ecommerce.business;
 
-import com.hhplus.ecommerce.domain.*;
-import com.hhplus.ecommerce.infrastructure.BasketDetailRepository;
+import com.hhplus.ecommerce.domain.Basket;
+import com.hhplus.ecommerce.domain.BasketDetail;
+import com.hhplus.ecommerce.domain.Item;
+import com.hhplus.ecommerce.domain.User;
+import com.hhplus.ecommerce.infrastructure.BasketRepository;
+import com.hhplus.ecommerce.infrastructure.ItemRepository;
+import com.hhplus.ecommerce.infrastructure.UserRepository;
+import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-
-@ExtendWith(MockitoExtension.class)
+@SpringBootTest
+@Transactional
 class BasketDetailServiceTest {
 
-    @Mock
-    private BasketDetailRepository basketDetailRepository;
-
-    @InjectMocks
+    @Autowired
     private BasketDetailService basketDetailService;
 
-    private User mockUser;
-    private Basket mockBasket;
-    private Item mockItem1;
-    private Item mockItem2;
+    @Autowired
+    private BasketRepository basketRepository;
+
+    @Autowired
+    private ItemRepository itemRepository;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    private Basket saveBasket;
+    private Item saveItem1;
+    private Item saveItem2;
+    private User saveUser;
+
+    @Autowired
+    private BasketService basketService;
 
     @BeforeEach
     void setUp() {
-        reset(basketDetailRepository);
-        mockUser = new User("김태현", 0, LocalDateTime.now());
-        mockBasket = new Basket(mockUser, LocalDateTime.now());
-        mockItem1 = new Item("청바지1", 10000, 10, LocalDateTime.now());
-        mockItem2 = new Item("청바지2", 20000, 20, LocalDateTime.now());
+        saveUser = userRepository.save(new User("김태현", 0, LocalDateTime.now()));
+        saveItem1 = itemRepository.save(new Item("청바지A", 9999, 4, LocalDateTime.now()));
+        saveItem2 = itemRepository.save(new Item("청바지B", 9999, 2, LocalDateTime.now()));
     }
 
     @Test
-    @DisplayName("상품_장바구니에_담기")
-    void saveBasketDetail() {
-        //given
-        BasketDetail mockBasketDetail = new BasketDetail(mockBasket, mockItem1, 1);
-        when(basketDetailRepository.save(mockBasketDetail)).thenReturn(mockBasketDetail);
-        //when
-        BasketDetail resultBasketDetail = basketDetailService.save(mockBasketDetail);
+    @DisplayName("상품_추가")
+    void addItemInBasketDetail() {
+        // given
+        Basket basket = new Basket(saveUser, LocalDateTime.now());
+        List<BasketDetail> basketDetails = new ArrayList<>();
+        BasketDetail basketDetail = new BasketDetail(basket, saveItem1, 2);
+        basketDetails.add(basketDetail);
+        for (BasketDetail detail : basketDetails) {
+            basket.addBasketDetail(detail);
+        }
+        // when
+        Basket saveBasket = basketRepository.save(basket);
         //then
-        assertNotNull(resultBasketDetail);
-        assertEquals(mockBasketDetail.getBasketDetailId(), resultBasketDetail.getBasketDetailId());
+        assertNotNull(saveBasket);
+        assertEquals(saveBasket.getBasketDetails().size(), basketDetails.size());
     }
 
     @Test
-    @DisplayName("상품_장바구니에서_제거")
-    void deleteBasketDetail() {
+    @DisplayName("상품_삭제")
+    void deleteItemInBasketDetail() {
         //given
-        BasketDetail mockBasketDetail = new BasketDetail(mockBasket, mockItem1, 1);
-        when(basketDetailRepository.findById(mockBasketDetail.getBasketDetailId())).thenReturn(Optional.of(mockBasketDetail));
+        Basket settingBasket = settingBasket();
+        Long basketDetailId = settingBasket.getBasketDetails().get(0).getBasketDetailId();
         //when
-        basketDetailService.delete(mockBasketDetail.getBasketDetailId());
+        BasketDetail deleteBasketDetail = basketDetailService.getBasketDetail(basketDetailId);
+        settingBasket.getBasketDetails().remove(deleteBasketDetail);
         //then
-        verify(basketDetailRepository, times(1)).delete(mockBasketDetail);
+        assertEquals(1, settingBasket.getBasketDetails().size());
     }
 
-    @Test
-    @DisplayName("장바구니에_아이템이_있는_확인")
-    void findItemInBasketDetail() {
-        //given
-        BasketDetail mockBasketDetail = new BasketDetail(mockBasket, mockItem1, 1);
-        when(basketDetailRepository.findBasketDetailByItem(mockItem1)).thenReturn(mockBasketDetail);
-        //when
-        BasketDetail resultBastDetail = basketDetailService.findItemInBasketDetail(mockItem1);
-        //then
-        assertNotNull(resultBastDetail);
-        assertEquals(mockBasketDetail.getBasketDetailId(), resultBastDetail.getBasketDetailId());
-    }
+    private Basket settingBasket() {
+        Basket basket = new Basket(saveUser, LocalDateTime.now());
+        List<BasketDetail> basketDetails = new ArrayList<>();
 
-    @Test
-    @DisplayName("장바구니번호로_상세_내역_전부_조회")
-    void getAllItemInBasketDetail() {
-        //given
-        BasketDetail mockBasketDetail1 = new BasketDetail(mockBasket, mockItem1, 1);
-        BasketDetail mockBasketDetail2 = new BasketDetail(mockBasket, mockItem2, 1);
-        List<BasketDetail> mockBasketDetailList = new ArrayList<>();
-        mockBasketDetailList.add(mockBasketDetail1);
-        mockBasketDetailList.add(mockBasketDetail2);
-        when(basketDetailRepository.findAllByBasket(mockBasket)).thenReturn(mockBasketDetailList);
-        //when
-        List<BasketDetail> resultBasketDetailList = basketDetailService.getAllDetailByBasket(mockBasket);
-        // then
-        assertNotNull(resultBasketDetailList);
-        assertEquals(mockBasketDetailList.size(), resultBasketDetailList.size());
+        BasketDetail basketDetail1 = new BasketDetail(basket, saveItem1, 2);
+        BasketDetail basketDetail2 = new BasketDetail(basket, saveItem2, 2);
+        basketDetails.add(basketDetail1);
+        basketDetails.add(basketDetail2);
+
+        for (BasketDetail detail : basketDetails) {
+            basket.addBasketDetail(detail);
+        }
+
+        return basketRepository.save(basket);
     }
 }
